pyproject.toml
[project]
name = "trendradar"
version = "5.4.0"
description = "TrendRadar - 热点新闻聚合与分析工具"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.5,<3.0.0",
    "pytz>=2025.2,<2026.0",
    "PyYAML>=6.0.3,<7.0.0",
    "fastmcp>=2.12.0,<2.14.0",
    "websockets>=13.0,<14.0",
    "feedparser>=6.0.0,<7.0.0",
    "boto3>=1.35.0,<2.0.0",
    "litellm>=1.57.0,<2.0.0",
    "tenacity==8.5.0",
    "opencc-python-reimplemented>=0.1.7,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "apscheduler>=3.10.0,<4.0.0"
]

[project.scripts]
trendradar = "trendradar.__main__:main"
trendradar-mcp = "mcp_server.server:run_server"

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trendradar", "mcp_server", "packages", "apps"]
